Fixing data type in startActivity and use equals()

+++ src/main/java/com/atixlabs/semillasmiddleware/excelparser/app/constants/EntrepreneurshipQuestion.java
@@ -23,7 +23,7 @@ public enum EntrepreneurshipQuestion implements CategoryQuestion {
     ACTIVITY_START_DATE("FECHA DE INICIO"){
         @Override
         public Class<?> getDataType() {
-            return Long.class;
+            return Integer.class;
         }
     },
     MAIN_ACTIVITY("ACTIVIDAD PRINCIPAL"),
